Move EnableTestMode to TasksManager
EnableTestMode is used to notify tests when a task run is completed.
TasksManager will be responsible for running a task vs. ConditionMonitor will
be responsible for triggering a task.

EnableTestMode should be in TasksManager.
